server/requirements.txt
aiohttp
aiopg
sanic >= 0.7
lxml ==4.1.1
asyncpg
sqlalchemy ==1.2.0b3
psycopg2
sqlalchemy-utils
requests
tqdm
sanic-cors
python-slugify